WIP floating rect abc
