--- src/Selection.css
@@ -1,14 +1,0 @@
-.Selection {
-  background: rgba(160, 195, 255, 0.2);
-  border: 2px #4285f4 solid;
-  position: absolute;
-  pointer-events: none;
-}
-
-.Selection.hidden {
-  display: none;
-}
-
-.Selection.copied {
-  border-style: dashed;
-}
